Move to config package
