+++ pkg/test/config/toolchainconfig.go
@@ -1,7 +1,8 @@
-package test
+package config
 
 import (
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
+	"github.com/codeready-toolchain/toolchain-common/pkg/test"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -137,7 +138,7 @@ func (o MembersOption) SpecificPerMemberCluster(clusterName string, memberConfig
 func NewToolchainConfig(options ...ToolchainConfigOption) *toolchainv1alpha1.ToolchainConfig {
 	toolchainConfig := &toolchainv1alpha1.ToolchainConfig{
 		ObjectMeta: metav1.ObjectMeta{
-			Namespace: HostOperatorNs,
+			Namespace: test.HostOperatorNs,
 			Name:      "config",
 		},
 	}
